+++ projects/ontimize-web-ngx/src/lib/components/table/index.ts
@@ -2,3 +2,4 @@ export * from './column/index';
 export * from './extensions/index';
 export * from './o-table.component';
 export * from './o-table.module';
+export * from './utils/o-table.tokens';